pyproject.toml
[project]
name = "jwb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ablog>=0.11.12",
    "aiokafka>=0.12.0",
    "matplotlib>=3.10.7",
    "myst-nb>=1.3.0",
    "pandas>=2.3.3",
    "pydata-sphinx-theme>=0.16.1",
    "redis>=7.0.0",
    "rstfmt>=0.0.14",
    "scipy>=1.16.2",
    "sphinx>=8.2.3",
    "sphinx-copybutton>=0.5.2",
    "sphinx-sitemap>=2.9.0",
    "sphinx-togglebutton>=0.3.2",
    "sphinxext-opengraph>=0.13.0",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "fastapi>=0.119.0",
    "pre-commit>=4.3.0",
]
